+++ run.sh
@@ -1,1 +1,1 @@
-docker run -p 6379:6379 -it --rm redislabs/redisgraph:latest
+docker run -p 6379:6379 -it --rm redis/redis-stack-server