+++ CHANGELOG.md
@@ -2,4 +2,7 @@
 
 ## v0.1.0
 
-* Initial release
+* Initial release of the library
+* Support for Cypher query building
+* Communication with RedisGraph
+* Representation of real data through Graph/Node/Relationhip structures.